requirements.txt
keras==2.8.0
numpy==1.22.3
opencv_python==4.5.5.64
pandas==1.4.2
Pillow==9.1.0
streamlit==1.8.1
tensorflow==2.8.0
